+++ sdk/aqueduct/artifacts/numeric_artifact.py
@@ -9,8 +9,8 @@ from aqueduct.artifacts import bool_artifact
 from aqueduct.artifacts import utils as artifact_utils
 from aqueduct.artifacts.base_artifact import BaseArtifact
 from aqueduct.artifacts.metadata import ArtifactMetadata
-from aqueduct.dag import (
-    DAG,
+from aqueduct.dag import DAG
+from aqueduct.dag_deltas import (
     AddOrReplaceOperatorDelta,
     RemoveCheckOperatorDelta,
     apply_deltas_to_dag,
