+++ sdk/aqueduct/operators.py
@@ -13,7 +13,8 @@ from aqueduct.enums import (
     OperatorType,
     S3TableFormat,
     SalesforceExtractType,
-    ServiceType, SerializationType,
+    SerializationType,
+    ServiceType,
 )
 from aqueduct.error import AqueductError, InvalidUserArgumentException
 from aqueduct.integrations.integration import IntegrationInfo
@@ -154,7 +155,6 @@ class CheckSpec(BaseModel):
 
 class ParamSpec(BaseModel):
     val: str
-    type: ArtifactType
     serialization_type: SerializationType
 
 
@@ -216,6 +216,7 @@ def get_operator_type_from_spec(spec: OperatorSpec) -> OperatorType:
         return OperatorType.SYSTEM_METRIC
     else:
         raise AqueductError("Invalid operator type")
+
 
 # TODO: remov ethis
 def serialize_parameter_value(name: str, val: Any) -> str:
