+++ src/golang/cmd/server/request/parameters.go
@@ -1,6 +1,7 @@
 package request
 
 import (
+	"encoding/base64"
 	"encoding/json"
 	"net/http"
 
@@ -13,7 +14,7 @@ const (
 )
 
 // Extracts the parameters dictionary from the request, which maps
-// parameter name to its json-serialized string value.
+// parameter name to its base64-encoded, bytes-serialized value.
 func ExtractParamsfromRequest(r *http.Request) (map[string]string, error) {
 	serializedParams := []byte(r.FormValue(parametersKey))
 
@@ -28,15 +29,12 @@ func ExtractParamsfromRequest(r *http.Request) (map[string]string, error) {
 		return nil, err
 	}
 
+	// Check that the parameter string is base-64 encoded.
 	for param_name, param_val := range params {
-		if !IsJSON(param_val) {
-			return nil, errors.Newf("The value %s provided for parameter %s is not in a valid json format.", param_val, param_name)
+		_, err = base64.StdEncoding.DecodeString(param_val)
+		if err != nil {
+			return nil, errors.Newf("Internal error: parameter values must be base64-encoded. The value %s provided for parameter %s was not.", param_val, param_name)
 		}
 	}
 	return params, nil
-}
-
-func IsJSON(str string) bool {
-	var js json.RawMessage
-	return json.Unmarshal([]byte(str), &js) == nil
 }
