+++ src/golang/lib/collections/operator/param/param.go
@@ -1,10 +1,7 @@
 package param
 
-import "github.com/aqueducthq/aqueduct/lib/collections/artifact"
-
 // The value of a parameter must be JSON serializable.
 type Param struct {
-	Val               string        `json:"val"`
-	Type              artifact.Type `json:"type"`
-	SerializationType string        `json:"serialization_type"`
+	Val               string `json:"val"`
+	SerializationType string `json:"serialization_type"`
 }
