+++ src/python/aqueduct_executor/operators/param_executor/spec.py
@@ -1,8 +1,7 @@
 import json
 
-from pydantic import BaseModel, parse_obj_as
-
 from aqueduct_executor.operators.utils.enums import ArtifactType, SerializationType
+from pydantic import BaseModel, parse_obj_as
 
 try:
     from typing import Literal
