+++ src/python/aqueduct_executor/operators/utils/enums.py
